[FEATURE] Support logging
`Migrator#apply` and `Migrator#rollback` takes a `:logger` option to
set up custom logger where results to be sent.

If no logger is provided, then default `ROM::Migrator::Logger` is used.

Notice, the logger is a mutable IO object, that's why it is not stored
as an attribute of the immutable migrator, but used as a method argument only.

== Added

* ROM::Migrator::Logger (default class for loggers)
* ROM::Migrator#apply (takes `:logger` option)
* ROM::Migrator#rollback (takes `:logger` option)